Extract sticky user message preview logic to new module and add more tests.
